fpgaperf: move Radiant and Icecube to separate files
